final UI changes to forms
