+++ components/admin/cms/permits/permits.tsx
@@ -8,9 +8,9 @@ import {
   CardFooter,
 } from "@/components/ui/card";
 import { useState } from "react";
-import GatePassForm from "./forms/gate-pass";
-import WorkPermitForm from "./forms/work-permit";
-import TempParkingForm from "./forms/temp-parking";
+import GatePassForm from "./forms/gate-pass/gate-pass";
+import WorkPermitForm from "./forms/work-permit/work-permit";
+import TempParkingForm from "./forms/temp-parking/temp-parking";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useRouter, useSearchParams } from "next/navigation";
 import React from "react";
