+++ components/global/sites.tsx
@@ -1,3 +1,5 @@
+
+
 interface Building {
     name: string;
     floors: number[];
