+++ components/global/workdata.tsx
@@ -1,0 +1,83 @@
+const workTypes = [
+    "Hot Works",
+    "Electrical",
+    "Mechanical",
+    "Civis/Structural Work",
+    "Confine Space",
+    "Lifting/Working on Heights",
+    "Pest Control",
+    "Event Set-up",
+    "IT Related Works",
+    "Line Breaking",
+    "Repair and Maintenance",
+    "Sanitary",
+    "Painting Works",
+    "Carpentry",
+    "Office Equipment/Furnitures/Others",
+    "Welding/Cutting",
+    "Installation",
+    "Delivery",
+    "Pick-up",
+    "Communications",
+    "Civil/Carpentry/Painting",
+    "Air Conditioning",
+    "Plumbing",
+    "Construction",
+    "Masonry",
+    "Renovation",
+    "Exhaust/Fresh Air",
+    "Maintenance",
+    "Gas Line",
+    "Repair",
+    "Sprinkler",
+    "Spray Painting",
+    "Sandling",
+    "Welding",
+    "Waterproofing",
+    "Draining/Refilling (Fire Protection)",
+    "Repair & Maintenance",
+    "Fire Protection",
+    "Sanding",
+    "Noise at workplace",
+    "Manual Handling",
+    "Fit-out Control",
+    "Work with Asbestos",
+    "Works with Pressurized System",
+    "Handling with Sharps",
+    "Handling with hazardous materials",
+    "Carpet Cleaning",
+    "Fiber Optic Cable Pulling",
+    "Others",
+    "Testing",
+    "Minor Works",
+  ];
+
+  const workRequirements = [
+    "Fire Extinguisher",
+    "Lockout/Tagout",
+    "Waste Segregation",
+    "Noise Control",
+    "First Aid Kit",
+    "Fire Blanket",
+    "Protective Cover",
+    "Fire Watcher/Watcher",
+    "PPE",
+    "Area Barricaded/Signages",
+    "Atmospheric testing",
+    "Housekeeping",
+    "Spill Control",
+    "Contractor Induction/Safety Briefing",
+    "List of Tools Equipment and Materials",
+    "Time Table (Gantt Chart)",
+    "Job Hazard Statement Form",
+    "Hot Work Permit",
+    "Fire Alarm Isolation Form",
+    "Plans/Layout Installation Details etc",
+    "Assistance Required",
+    "Electrical Circuit Breaker",
+    "Warning Tape",
+    "IT Equipments Installation",
+    "Aircon",
+  ];
+
+  export {workTypes, workRequirements};
